Refactoring

+++ LAB1/src/MatriceAddThread.java
@@ -4,10 +4,10 @@ public class MatriceAddThread implements Runnable {
     private Matrice a;
     private Matrice b;
     private Matrice c;
-    private int startIndex;
-    private int endIndex;
+    private long startIndex;
+    private long endIndex;
 
-    public MatriceAddThread(Matrice a, Matrice b, Matrice c, int startIndex, int endIndex) {
+    public MatriceAddThread(Matrice a, Matrice b, Matrice c, long startIndex, long endIndex) {
         this.a = a;
         this.b = b;
         this.c = c;
@@ -18,9 +18,9 @@ public class MatriceAddThread implements Runnable {
     @Override
     public void run() {
         final int columns = a.getColumns();
-        for (int i = startIndex; i < endIndex; i++) {
-            final int line = i / columns;
-            final int col = i % columns;
+        for (long i = startIndex; i < endIndex; i++) {
+            final int line = (int) (i / columns);
+            final int col = (int) (i % columns);
 //            System.out.println(line + " " + col);
             c.setElement(line, col, a.getElement(line, col) + b.getElement(line, col));
         }
